tidy things up a bit
